[feat] Add comprehensive system configuration management for QuBE
- Implemented `SystemConfigDatabase` class to manage system configurations, including clockmaster, boxes, ports, and channels.
- Added methods to define, set, and retrieve configurations for clockmaster, boxes, ports, and channels.
- Introduced serialization methods (`asdict`, `asjson`) for exporting configurations.
- Enhanced error handling for invalid configurations and missing targets.
- Included `ClockmasterSetting`, `BoxSetting`, and `PortSetting` dataclasses for structured configuration management.
- Integrated support for Quel1 system creation and box management.

This update provides a robust framework for managing and interacting with QuBE system configurations.
